requirements.txt
streamlit==1.25.0
PyPDF2==3.0.1
langchain_community==0.1.0
langchain_core==0.1.0
transformers==4.35.0
